\name{ozone}
\alias{ozone}
\docType{data}
\title{Ozone data from Breiman and Friedman, 1985}
\description{
This is the Ozone data discussed in Breiman and Friedman (JASA, 1985, p. 580).
These data are for 330 days in 1976.  All measurements are in the area of
Upland, CA, east of Los Angeles.
}
\format{This data frame contains the following columns:
 \describe{
 \item{Ozone:}{Ozone conc., ppm, at Sandbug AFB.}
 
\item{Temp:}{Temperature F. (max?).}

 \item{InvHt:}{Inversion base height, feet}

 \item{Pres:}{Daggett pressure gradient (mm Hg)}

 \item{Vis:}{Visibility (miles)}

 \item{Hgt:}{Vandenburg 500 millibar height (m)}
 \item{Hum:}{Humidity, percent}

 \item{InvTmp:}{Inversion base temperature, degrees F.}
 \item{Wind:}{Wind speed, mph}
}}


\usage{data(ozone)}
\source{Breiman, L and  Friedman, J.  (1985),
 ``Estimating Optimal Transformations for Multiple Regression and Correlation'',
 \emph{Journal of the American Statistical Association}, 80,  580-598.
    }
\keyword{datasets}
