\name{wine}
\alias{wine}
\docType{data}
\title{Wine recognition data}
\description{
Data from the machine learning repository.
A chemical analysis
of 178 Italian wines from three different cultivars yielded 13  measurements.
This dataset is often used to test and compare the performance of various
classification algorithms.

}
\format{This data frame contains the following columns:
 \describe{
 \item{Class:}{There are 3 classes}
 \item{Alcohol:}{Alcohol}
 \item{Malic:}{Malic acid}
 \item{Ash:}{Ash}
 \item{Alcalinity:}{Alcalinity of ash}
 \item{Magnesium:}{Magnesium}
 \item{Phenols:}{Total phenols}
 \item{Flavanoids:}{Flavanoids}
 \item{Nonflavanoid:}{Nonflavanoid phenols}
 \item{Proanthocyanins:}{Proanthocyanins}
 \item{Intensity:}{Color intensity}
 \item{Hue:}{Hue}
 \item{OD280:}{OD280/OD315 of diluted wines}
 \item{Proline:}{Proline}
}}
\usage{data(wine)}


\source{
Forina, M. et al, PARVUS - An Extendible Package for Data
       Exploration, Classification and Correlation. Institute of Pharmaceutical
       and Food Analysis and Technologies, Via Brigata Salerno, 
       16147 Genoa, Italy.
}
\references{
Blake, C.L. and Merz, C.J. (1998), UCI Repository of machine learning databases, \\
\url{https://archive.ics.uci.edu/datasets}. Irvine, CA: University of California, Department of
Information and Computer Science. 
}
\keyword{datasets}
